+++ core/src/test/scala/com/colisweb/jrubysnesshours/core/ScheduleWithCutOffSpec.scala
@@ -1,9 +1,10 @@
 package com.colisweb.jrubysnesshours.core
 
-import com.colisweb.Approbation
 import com.colisweb.jrubysnesshours.core.SpecUtils._
+import com.github.writethemfirst.Approbation
+import org.scalatest.{Matchers, fixture}
 
-class ScheduleWithCutOffSpec extends Approbation {
+class ScheduleWithCutOffSpec extends fixture.FlatSpec with Matchers with Approbation {
   val cutOff = Some(
     DoubleCutOff(
       sameDay = CutOff(limit = "08:00".toLocalTime, firstAvailableTime = "17:00".toLocalTime),
@@ -31,7 +32,7 @@ class ScheduleWithCutOffSpec extends Approbation {
     approver.verify(prettify(slots))
   }
 
-  it should "cut start of D+1 day" in { approver =>
+  it should "cut start of D plus 1 day" in { approver =>
     val slots = schedule.splitTimeSegments(
       "2019-05-02" at "12:01" -> FRANCE_TIMEZONE,
       "2019-05-12" at "10:00" -> FRANCE_TIMEZONE,